+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/CSharp/Extensions/MemberAccessExpressionSyntaxExtensions.cs
