+++ src/Workspaces/SharedUtilitiesAndExtensions/Compiler/CSharp/Extensions/SyntaxNodeExtensions.cs
@@ -875,4 +875,23 @@ internal static partial class SyntaxNodeExtensions
         => node
         .WithTrailingTrivia(node.GetTrailingTrivia().FilterComments(addElasticMarker: true))
         .WithLeadingTrivia(node.GetLeadingTrivia().FilterComments(addElasticMarker: true));
+
+    public static SyntaxNode WithPrependedNonIndentationTriviaFrom(
+        this SyntaxNode to, SyntaxNode from)
+    {
+        // get all the preceding trivia from the 'from' node, not counting the leading
+        // indentation trivia is has.
+        var finalTrivia = from.GetLeadingTrivia().ToList();
+        while (finalTrivia is [.., (kind: SyntaxKind.WhitespaceTrivia)])
+            finalTrivia.RemoveAt(finalTrivia.Count - 1);
+
+        // Also, add on the trailing trivia if there are trailing comments.
+        var hasTrailingComments = from.GetTrailingTrivia().Any(t => t.IsRegularComment());
+        if (hasTrailingComments)
+            finalTrivia.AddRange(from.GetTrailingTrivia());
+
+        // Merge this trivia with the existing trivia on the node.  Format in case
+        // we added comments and need them indented properly.
+        return to.WithPrependedLeadingTrivia(finalTrivia);
+    }
 }
